the New God/requirements.txt
streamlit
streamlit_webrtc
Pillow
opencv-python-headless
numpy
tensorflow
keras
keras-vggface
cvlib
